api improvements, rabbitmq memleaks fix, support cadvisor metrics, deb package support

+++ src/events/fs_write.h
@@ -1,2 +1,2 @@
 #pragma once
-void write_to_file(char *filename, char *str, size_t len, void *callback, void *data);
+void write_to_file(char *filename, char *str, uint64_t len, void *callback, void *data);
